Moves logging into its own module
